requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-dotenv==1.0.1
SQLAlchemy==2.0.28
psycopg2==2.9.9
python-multipart==0.0.9
python-jose==3.3.0
bcrypt==4.1.2
email-validator==2.1.1
jinja2==3.1.3